refactor(dashboard): consolidate overlapping pages — delete a dead AI cluster, merge two config pages
Two separate kinds of bloat, measured rather than guessed at.

1. A dead AI cluster (1,109 lines deleted)

`ai-analysis` (638), `ai-governance` (220) and `ai-monitoring` (251) were
unreachable: absent from the sidebar, and the `ai-security` hub that aggregates
their stats does not link to them either. The only inbound link was
ai-analysis -> ai-governance, and ai-analysis is itself orphaned, so the three
formed a closed unreachable cluster. All read-only — zero write calls, zero form
inputs — so nothing but detail views is lost.

ControlWeaver-Pro had already deleted exactly these three and kept `ai-security`
as the consolidated hub. This repo never received that change, so it had been
carrying them as dead weight. Now at parity: ai-insights, ai-laws, ai-security.

2. Departments + Business Objectives merged into one page

Both were thin org-configuration lists (341 and 333 lines) — a list, a create
form, a roll-up, no lifecycle of their own — and two top-level nav entries for
that is not worth it. They are also read together: you assign an objective to a
department, and a department's open-risk count only means something next to the
objectives that department owns.

New `/dashboard/structure` hosts both as tabs, with the bodies extracted to
`components/structure/{Departments,Objectives}Panel.tsx`. Each tab is
permission-gated independently, and the tab actually shown falls back to the
first one the user can see — so someone holding only `objectives.read` gets the
Objectives tab rather than a blank Departments view.

`/dashboard/departments` and `/dashboard/objectives` become redirect stubs to the
matching tab rather than 404s, following the pattern `dashboard/audit` already
uses after folding into settings. Bookmarks and the docs keep working.

Sidebar drops from 48 entries to 47: two entries become one "Structure" under
Organization, gated on either permission.

Deliberately left alone, with reasons:

- vendor-risk (609) and tprm (1,984) overlap conceptually but merging them makes
  a 2,600-line page; the overlap is in the domain, not the code.
- security-posture / threat-intel / plot4ai / sbom / vulnerabilities read as one
  cluster but draw on five different data sources, and vulnerabilities alone is
  1,183 lines.
- my-organizations (221) and organization (1,119) sound alike but do different
  jobs — switching between orgs versus editing the current one's profile.

Verified live, not just by build: /dashboard/departments and
/dashboard/objectives both redirect to the correct tab, both tabs render their
seeded data (8 departments, 6 objectives), typecheck clean, production build 97
pages.
